Disable implicit casting between uint128 and int128; add more tests; fix whitespace error in Makefile; some refactoring

+++ src/include/function/cast/functions/cast_string_non_nested_functions.h
@@ -126,7 +126,7 @@ struct Int128CastData {
             if (digits > DECIMAL_PRECISION_LIMIT) {
                 return false;
             }
-            if (!Int128_t::tryMultiply(result, UInt128_t::powerOf10[digits], result)) {
+            if (!Int128_t::tryMultiply(result, (int128_t)UInt128_t::powerOf10[digits], result)) {
                 return false;
             }
         }
